Update circuit.R
Fixed some logical errors. The answer needs to be multiplied by (m-n+1)/(m+1) since the length of the travel time vector is m-n+1. Furthermore the interchange happening with the help of the replace function was using the values for the index vector not the indexes. Works only if the sequence is 1:20. Finally when we assign forbidden pairs (i,j) to our distance matrix, it must be in the form of D[i,j] where i<j.